Fix compilation problem with CbmRoot
The compilation of some CbmRoot analysis files did crash due to missing
header files when not setting the environment variables SIMPATH and
FAIRROOTPATH.
The problem is that CbmRoot allows to pass this information as CMake
parameters without setting the environment variables but the info
is expected as environment variables by Hal. If the environment variable
FAIRROOTPATH isn't found not all targets are compiled and installed
which in the end result in crashes during CbmRoot compilation.